requirements.txt
torch
google-cloud-texttospeech
pydub
soundfile
requests
